requirements.txt
alabaster==0.7.12
amqp==2.5.2
argh==0.26.2
Babel==2.8.0
billiard==3.6.1.0
boto3==1.10.46
botocore==1.13.46
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
Django==2.2.7
django-extensions==2.2.5
djangorestframework==3.10.3
docutils==0.15.2
futures==3.1.1
idna==2.8
imageio==2.6.1
imagesize==1.2.0
Jinja2==2.10.3
jmespath==0.9.4
kiwisolver==1.1.0
kombu==4.6.7
MarkupSafe==1.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.18.0
packaging==19.2
pathtools==0.1.2
Pillow==7.0.0
pipdeptree==0.13.2
psycopg2==2.8.4
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
redis==3.3.11
requests==2.22.0
s3transfer==0.2.1
scikit-image==0.16.2
scipy==1.4.1
six==1.13.0
snowballstemmer==2.0.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
watchdog==0.9.0
xmltodict==0.12.0
